remove local bus struct
removes the local bus from fault in favor of a single Bus that
operates both within a local and global instance, and can be
passed downstream independent of context.  Also includes
some code separation in the fault package for readability.
