+++ src/pkg/fault/item.go
@@ -11,15 +11,15 @@ const (
 	AddtlMalwareDesc   = "malware_description"
 )
 
-type itemType string
+type ItemType string
 
 const (
-	FileType          itemType = "file"
-	ContainerType     itemType = "container"
-	ResourceOwnerType itemType = "resource_owner"
+	FileType          ItemType = "file"
+	ContainerType     ItemType = "container"
+	ResourceOwnerType ItemType = "resource_owner"
 )
 
-func (it itemType) Printable() string {
+func (it ItemType) Printable() string {
 	switch it {
 	case FileType:
 		return "File"
@@ -62,7 +62,7 @@ type Item struct {
 	Name string `json:"name"`
 
 	// tracks the type of item represented by this entry.
-	Type itemType `json:"type"`
+	Type ItemType `json:"type"`
 
 	// Error() of the causal error, or a sentinel if this is the
 	// source of the error.  In case of ID collisions, the first
@@ -138,7 +138,7 @@ func OwnerErr(cause error, namespace, id, name string, addtl map[string]any) *It
 }
 
 // itemErr produces a Item of the provided type for tracking erroneous items.
-func itemErr(t itemType, cause error, namespace, id, name string, addtl map[string]any) *Item {
+func itemErr(t ItemType, cause error, namespace, id, name string, addtl map[string]any) *Item {
 	return &Item{
 		Namespace:  namespace,
 		ID:         id,
@@ -146,185 +146,5 @@ func itemErr(t itemType, cause error, namespace, id, name string, addtl map[stri
 		Type:       t,
 		Cause:      cause.Error(),
 		Additional: addtl,
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Skipped Items
-// ---------------------------------------------------------------------------
-
-// skipCause identifies the well-known conditions to Skip an item.  It is
-// important that skip cause enumerations do not overlap with general error
-// handling.  Skips must be well known, well documented, and consistent.
-// Transient failures, undocumented or unknown conditions, and arbitrary
-// handling should never produce a skipped item. Those cases should get
-// handled as normal errors.
-type skipCause string
-
-const (
-	// SkipMalware identifies a malware detection case.  Files that graph
-	// api identifies as malware cannot be downloaded or uploaded, and will
-	// permanently fail any attempts to backup or restore.
-	SkipMalware skipCause = "malware_detected"
-
-	// SkipBigOneNote identifies that a file was skipped because it
-	// was big OneNote file and we can only download OneNote files which
-	// are less that 2GB in size.
-	//nolint:lll
-	// https://support.microsoft.com/en-us/office/restrictions-and-limitations-in-onedrive-and-sharepoint-64883a5d-228e-48f5-b3d2-eb39e07630fa#onenotenotebooks
-	SkipBigOneNote skipCause = "big_one_note_file"
-)
-
-var _ print.Printable = &Skipped{}
-
-// Skipped items are permanently unprocessable due to well-known conditions.
-// In order to skip an item, the following conditions should be met:
-// 1. The conditions for skipping the item are well-known and
-// well-documented.  End users need to be able to understand
-// both the conditions and identifications of skips.
-// 2. Skipping avoids a permanent and consistent failure.  If
-// the underlying reason is transient or otherwise recoverable,
-// the item should not be skipped.
-//
-// Skipped wraps Item primarily to minimize confusion when sharing the
-// fault interface.  Skipped items are not errors, and Item{} errors are
-// not the basis for a Skip.
-type Skipped struct {
-	Item Item `json:"item"`
-}
-
-// String complies with the stringer interface.
-func (s *Skipped) String() string {
-	if s == nil {
-		return "<nil>"
-	}
-
-	return "skipped " + s.Item.Error() + ": " + s.Item.Cause
-}
-
-// HasCause compares the underlying cause against the parameter.
-func (s *Skipped) HasCause(c skipCause) bool {
-	if s == nil {
-		return false
-	}
-
-	return s.Item.Cause == string(c)
-}
-
-func (s Skipped) MinimumPrintable() any {
-	return s
-}
-
-// Headers returns the human-readable names of properties of a skipped Item
-// for printing out to a terminal.
-func (s Skipped) Headers() []string {
-	return []string{"Action", "Type", "Name", "Container", "Cause"}
-}
-
-// Values populates the printable values matching the Headers list.
-func (s Skipped) Values() []string {
-	var cn string
-
-	acn, ok := s.Item.Additional[AddtlContainerName]
-	if ok {
-		str, ok := acn.(string)
-		if ok {
-			cn = str
-		}
-	}
-
-	return []string{"Skip", s.Item.Type.Printable(), s.Item.Name, cn, s.Item.Cause}
-}
-
-// ContainerSkip produces a Container-kind Item for tracking skipped items.
-func ContainerSkip(cause skipCause, namespace, id, name string, addtl map[string]any) *Skipped {
-	return itemSkip(ContainerType, cause, namespace, id, name, addtl)
-}
-
-// FileSkip produces a File-kind Item for tracking skipped items.
-func FileSkip(cause skipCause, namespace, id, name string, addtl map[string]any) *Skipped {
-	return itemSkip(FileType, cause, namespace, id, name, addtl)
-}
-
-// OnwerSkip produces a ResourceOwner-kind Item for tracking skipped items.
-func OwnerSkip(cause skipCause, namespace, id, name string, addtl map[string]any) *Skipped {
-	return itemSkip(ResourceOwnerType, cause, namespace, id, name, addtl)
-}
-
-// itemSkip produces a Item of the provided type for tracking skipped items.
-func itemSkip(t itemType, cause skipCause, namespace, id, name string, addtl map[string]any) *Skipped {
-	return &Skipped{
-		Item: Item{
-			Namespace:  namespace,
-			ID:         id,
-			Name:       name,
-			Type:       t,
-			Cause:      string(cause),
-			Additional: addtl,
-		},
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Alerts
-// ---------------------------------------------------------------------------
-
-var _ print.Printable = &Alert{}
-
-// Alerts are informational-only notifications.  The purpose of alerts is to
-// provide a means of end-user communication about important events without
-// needing to generate runtime failures or recoverable errors. When generating
-// an alert, no other fault feature (failure, recoverable, skip, etc) should
-// be in use.  IE: Errors do not also get alerts, since the error itself is a
-// form of end-user communication already.
-type Alert struct {
-	Item    Item   `json:"item"`
-	Message string `json:"message"`
-}
-
-// String complies with the stringer interface.
-func (a Alert) String() string {
-	msg := a.Message
-	if len(msg) == 0 {
-		msg = "<nil>"
-	}
-
-	return "Alert: " + msg
-}
-
-func (a Alert) MinimumPrintable() any {
-	return a
-}
-
-// Headers returns the human-readable names of properties of a skipped Item
-// for printing out to a terminal.
-func (a Alert) Headers() []string {
-	return []string{"Action", "Message", "Container", "Name", "ID"}
-}
-
-// Values populates the printable values matching the Headers list.
-func (a Alert) Values() []string {
-	var cn string
-
-	acn, ok := a.Item.Additional[AddtlContainerName]
-	if ok {
-		str, ok := acn.(string)
-		if ok {
-			cn = str
-		}
-	}
-
-	return []string{"Alert", a.Message, cn, a.Item.Name, a.Item.ID}
-}
-
-func NewAlert(message, namespace, itemID, name string, addtl map[string]any) *Alert {
-	return &Alert{
-		Message: message,
-		Item: Item{
-			Namespace:  namespace,
-			ID:         itemID,
-			Name:       name,
-			Additional: addtl,
-		},
 	}
 }
